+++ app/users/repository/user_repository.py
@@ -1,5 +1,6 @@
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
+from app.users.exceptions import UserNotFoundException
 from app.users.models import User
 
 
@@ -35,10 +36,18 @@ class UserRepository:
 
     def get_user_by_id(self, user_id: str):
         user = self.db.query(User).filter(User.id == user_id).first()
+        if user is None:
+            raise UserNotFoundException(
+                f"User with provided id: {user_id} not found.", 400
+            )
         return user
 
     def get_user_by_email(self, email: str):
         user = self.db.query(User).filter(User.email == email).first()
+        if user is None:
+            raise UserNotFoundException(
+                f"User with provided email: {email} not found.", 400
+            )
         return user
 
     def get_all_users(self):
@@ -48,6 +57,10 @@ class UserRepository:
     def delete_user_by_id(self, user_id: str):
         try:
             user = self.db.query(User).filter(User.id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(
+                    f"User with provided id: {user_id} not found.", 400
+                )
             self.db.delete(user)
             self.db.commit()
             return True
@@ -57,6 +70,10 @@ class UserRepository:
     def update_user_name(self, user_id: str, name: str):
         try:
             user = self.db.query(User).filter(User.id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(
+                    f"User with provided id: {user_id} not found.", 400
+                )
             user.name = name
             self.db.add(user)
             self.db.commit()
@@ -68,6 +85,10 @@ class UserRepository:
     def update_user_surname(self, user_id: str, surname: str):
         try:
             user = self.db.query(User).filter(User.id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(
+                    f"User with provided id: {user_id} not found.", 400
+                )
             user.surname = surname
             self.db.add(user)
             self.db.commit()
@@ -79,6 +100,10 @@ class UserRepository:
     def update_user_is_active(self, user_id: str, is_active: bool):
         try:
             user = self.db.query(User).filter(User.id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(
+                    f"User with provided id: {user_id} not found.", 400
+                )
             user.is_active = is_active
             self.db.add(user)
             self.db.commit()
